requirements.txt
openai
faiss-cpu
gradio
numpy
